added ability to pick up stuff faster
Former-commit-id: 6d0aefa5bcbbe32b54dcddf817e5f74231b93713
Former-commit-id: 2b458b227ed7ba99e9a65dc6660fcee3c9306934

+++ Assets/Scripts/UI/UISaveGame.cs
@@ -221,7 +221,7 @@ public class UISaveGame : MonoBehaviour
         selectedSaveItem = uiSaveItem;
     }
 
-    private string ConvertPlayTime(float time)
+    private string ConvertPlayTime(int time)
     {
         return Math.Round(time * 0.0002777f, 1).ToString() + " hrs";
     }
@@ -348,7 +348,7 @@ public class UISaveGame : MonoBehaviour
         world.StartSaveProcess(saveName);
     }
 
-    public void UpdateSaveItems(string saveName, float savePlayTime, string saveVersion, int seed/*, Texture2D saveScreenshot*/)
+    public void UpdateSaveItems(string saveName, int savePlayTime, string saveVersion, int seed/*, Texture2D saveScreenshot*/)
     {
         if (newItem)
         {
